requirements.txt
# 多Agent协作系统依赖
# 安装命令: pip install -r requirements.txt

# Web框架
pywebview>=5.0.0

# 其他依赖（如果需要）
# requests>=2.28.0
# aiohttp>=3.8.0
